# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Lists and Tasks API"
  version: "1.0.0"
  description: "API для управления списками и задачами (CRUD, in-memory)"
servers:
  - url: http://localhost:8080
tags:
  - name: Lists
    description: "Операции со списками"
  - name: Tasks
    description: "Операции с задачами"
  - name: Health
    description: "Проверка состояния сервиса"
paths:
  /api/v1/lists:
    post:
      tags: [Lists]
      operationId: createList
      summary: "Создать список"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListRequest'
            examples:
              example:
                value:
                  title: Дом
      responses:
        '201':
          description: "Создано"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
    get:
      tags: [Lists]
      operationId: listLists
      summary: "Получить списки"
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: "Ок"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/List'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/lists/{id}:
    get:
      tags: [Lists]
      operationId: getList
      summary: "Получить список по id"
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: "Ок"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      tags: [Lists]
      operationId: updateList
      summary: "Обновить title списка"
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListRequest'
            examples:
              example:
                value:
                  title: Домашние дела
      responses:
        '200':
          description: "Ок"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Lists]
      operationId: deleteList
      summary: "Удалить список"
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: "Удалено"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/v1/lists/{listID}/tasks:
    post:
      tags: [Tasks]
      operationId: createTask
      summary: "Создать задачу в списке"
      parameters:
        - name: "listID"
          in: "path"
          description: "ID списка, к которому добавляется задача"
          required: true
          schema:
            type: string
      requestBody:
        description: "Информация о новой задаче"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        '201':
          description: "Задача успешно создана"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: "Некорректный запрос"
        '404':
          description: "Список не найден"
    get:
      tags: [Tasks]
      operationId: getTasks
      summary: "Получить все задачи в списке"
      parameters:
        - name: "listID"
          in: "path"
          description: "ID списка для получения задач"
          required: true
          schema:
            type: string
        - name: "limit"
          in: "query"
          description: "Количество задач для получения"
          required: false
          schema:
            type: integer
        - name: "offset"
          in: "query"
          description: "Смещение задач"
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: "Список задач"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        '404':
          description: "Список не найден"

  /api/v1/tasks/{taskID}:
    get:
      tags: [Tasks]
      operationId: getTask
      summary: "Получить задачу по ID"
      parameters:
        - name: "taskID"
          in: "path"
          description: "ID задачи"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Задача успешно найдена"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '404':
          description: "Задача не найдена"
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: "Обновить задачу"
      parameters:
        - name: "taskID"
          in: "path"
          description: "ID задачи для обновления"
          required: true
          schema:
            type: string
      requestBody:
        description: "Данные для обновления задачи"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        '200':
          description: "Задача обновлена"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        '400':
          description: "Некорректный запрос"
        '404':
          description: "Задача не найдена"
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: "Удалить задачу"
      parameters:
        - name: "taskID"
          in: "path"
          description: "ID задачи для удаления"
          required: true
          schema:
            type: string
      responses:
        '204':
          description: "Задача успешно удалена"
        '404':
          description: "Задача не найдена"

components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: UUID списка
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      required: false
      description: Максимум элементов в ответе
      schema:
        type: integer
        minimum: 0
    Offset:
      name: offset
      in: query
      required: false
      description: Смещение в выборке
      schema:
        type: integer
        minimum: 0

  schemas:
    List:
      type: object
      required: [id, title, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Идентификатор списка
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: Название списка
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Время создания (RFC3339)
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        title: Дом
        created_at: "2025-01-01T12:00:00Z"

    CreateListRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100

    UpdateListRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100

    Task:
      type: object
      required: [id, list_id, text, created_at, updated_at]
      properties:
        id:
          type: string
        list_id:
          type: string
        text:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500

    UpdateTaskRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
        completed:
          type: boolean

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
      example:
        code: "VALIDATION_FAILED"
        message: "title must be 1..100 chars"

    Health:
      type: object
      properties:
        status:
          type: string
          example: "ok"

  responses:
    NotFound:
      description: "Не найдено"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: "Ошибка валидации"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: "Ошибка сервера"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"